Move action registry to corecel/sys
