+++ app/celer-g4/GeantDiagnostics.cc
@@ -10,13 +10,13 @@
 #include "corecel/io/BuildOutput.hh"
 #include "corecel/io/Logger.hh"
 #include "corecel/io/OutputInterfaceAdapter.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "corecel/sys/Environment.hh"
 #include "corecel/sys/EnvironmentIO.json.hh"
 #include "corecel/sys/MemRegistry.hh"
 #include "corecel/sys/MemRegistryIO.json.hh"
 #include "corecel/sys/MultiExceptionHandler.hh"
 #include "celeritas/Types.hh"
-#include "celeritas/global/ActionRegistry.hh"
 #include "celeritas/global/CoreParams.hh"
 #include "celeritas/user/StepDiagnostic.hh"
