+++ src/celeritas/CMakeLists.txt
@@ -36,9 +36,6 @@ list(APPEND SOURCES
   field/RZMapFieldInputIO.json.cc
   field/RZMapFieldParams.cc
   geo/GeoMaterialParams.cc
-  global/ActionInterface.cc
-  global/ActionRegistry.cc
-  global/ActionRegistryOutput.cc
   global/CoreParams.cc
   global/CoreState.cc
   global/CoreTrackData.cc
