+++ src/celeritas/global/CoreState.cc
@@ -10,10 +10,10 @@
 #include "corecel/data/CollectionAlgorithms.hh"
 #include "corecel/data/Copier.hh"
 #include "corecel/io/Logger.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "corecel/sys/ScopedProfiling.hh"
 #include "celeritas/track/TrackInitParams.hh"
 
-#include "ActionRegistry.hh"
 #include "CoreParams.hh"
 
 namespace celeritas
