+++ src/celeritas/track/SortTracksAction.cc
@@ -12,8 +12,8 @@
 
 #include "corecel/Assert.hh"
 #include "corecel/Macros.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "celeritas/Types.hh"
-#include "celeritas/global/ActionRegistry.hh"
 #include "celeritas/global/CoreParams.hh"
 #include "celeritas/global/CoreState.hh"
