+++ src/celeritas/user/StepCollector.cc
@@ -16,8 +16,8 @@
 #include "corecel/data/CollectionBuilder.hh"
 #include "corecel/data/CollectionMirror.hh"
 #include "corecel/io/Label.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "celeritas/geo/GeoParams.hh"  // IWYU pragma: keep
-#include "celeritas/global/ActionRegistry.hh"
 
 #include "StepInterface.hh"
