+++ src/corecel/sys/ActionRegistry.hh
@@ -3,8 +3,7 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/global/ActionRegistry.hh
-//! \todo Move to corecel
+//! \file corecel/sys/ActionRegistry.hh
 //---------------------------------------------------------------------------//
 #pragma once
