+++ src/corecel/sys/ActionRegistryOutput.hh
@@ -3,8 +3,7 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/global/ActionRegistryOutput.hh
-//! \todo Move to corecel
+//! \file corecel/sys/ActionRegistryOutput.hh
 //---------------------------------------------------------------------------//
 #pragma once
