+++ test/celeritas/GlobalTestBase.cc
@@ -19,8 +19,8 @@
 #include "corecel/io/JsonPimpl.hh"
 #include "corecel/io/Logger.hh"
 #include "corecel/io/OutputRegistry.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "celeritas/ext/ScopedRootErrorHandler.hh"
-#include "celeritas/global/ActionRegistry.hh"
 #include "celeritas/global/CoreParams.hh"
 #include "celeritas/random/RngParams.hh"
 #include "celeritas/track/StatusChecker.hh"
