+++ test/celeritas/MockTestBase.cc
@@ -8,8 +8,8 @@
 #include "MockTestBase.hh"
 
 #include "corecel/math/Algorithms.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "celeritas/geo/GeoMaterialParams.hh"
-#include "celeritas/global/ActionRegistry.hh"
 #include "celeritas/global/alongstep/AlongStepGeneralLinearAction.hh"
 #include "celeritas/mat/MaterialParams.hh"
 #include "celeritas/phys/CutoffParams.hh"
