+++ test/celeritas/global/Stepper.test.cc
@@ -17,8 +17,8 @@
 #include "corecel/data/AuxParamsRegistry.hh"
 #include "corecel/io/LogContextException.hh"
 #include "corecel/io/Logger.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "geocel/UnitUtils.hh"
-#include "celeritas/global/ActionRegistry.hh"
 #include "celeritas/global/CoreParams.hh"
 #include "celeritas/global/CoreState.hh"
 #include "celeritas/global/alongstep/AlongStepUniformMscAction.hh"
