+++ test/celeritas/optical/OpticalCollector.test.cc
@@ -15,9 +15,9 @@
 #include "corecel/cont/Span.hh"
 #include "corecel/data/Copier.hh"
 #include "corecel/io/LogContextException.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "geocel/UnitUtils.hh"
 #include "celeritas/em/params/UrbanMscParams.hh"
-#include "celeritas/global/ActionRegistry.hh"
 #include "celeritas/global/Stepper.hh"
 #include "celeritas/global/alongstep/AlongStepUniformMscAction.hh"
 #include "celeritas/optical/detail/OffloadParams.hh"
