+++ test/celeritas/track/StatusChecker.test.cc
@@ -11,8 +11,8 @@
 #include "corecel/data/AuxParamsRegistry.hh"
 #include "corecel/data/CollectionBuilder.hh"
 #include "corecel/io/Logger.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "celeritas/SimpleTestBase.hh"
-#include "celeritas/global/ActionRegistry.hh"
 #include "celeritas/global/CoreParams.hh"
 #include "celeritas/global/CoreState.hh"
 #include "celeritas/track/SimTrackView.hh"
