+++ test/celeritas/track/TrackInit.test.cc
@@ -14,8 +14,8 @@
 #include "corecel/data/CollectionStateStore.hh"
 #include "corecel/data/Ref.hh"
 #include "corecel/io/LogContextException.hh"
+#include "corecel/sys/ActionRegistry.hh"
 #include "celeritas/SimpleTestBase.hh"
-#include "celeritas/global/ActionRegistry.hh"
 #include "celeritas/global/CoreParams.hh"
 #include "celeritas/global/CoreState.hh"
 #include "celeritas/global/CoreTrackData.hh"
