+++ test/corecel/CMakeLists.txt
@@ -76,6 +76,7 @@ celeritas_add_test(math/BisectionRootFinder.test.cc)
 celeritas_add_test(math/IllinoisRootFinder.test.cc)
 
 # Sys
+celeritas_add_test(sys/ActionRegistry.test.cc)
 celeritas_add_test(sys/Environment.test.cc
   ENVIRONMENT "ENVTEST_ONE=1;ENVTEST_ZERO=0;ENVTEST_EMPTY="
   LINK_LIBRARIES nlohmann_json::nlohmann_json
